Refactor navigation and adjust serialization settings
- Move navigation menu item creation to code-behind
- Replace DeserializerBuilder with optimized StaticDeserializerBuilder
- Remove unused rd.xml file and add link.xml for linker
- Move PublishAot, PublishTrimmed, SelfContained to separate PropertyGroup
- Disable EnableUnsafeBinaryFormatterSerialization in the project file

This change refactors the navigation menu item creation in MainWindow.axaml.cs to make it more maintainable. It also optimizes the DeserializerBuilder setup in Util.cs, replacing it with StaticDeserializerBuilder to improve performance. The commit removes an unused rd.xml file and adds a link.xml file for the linker, and reorganizes some project settings related to publishing for better readability. Lastly, it disables the EnableUnsafeBinaryFormatterSerialization setting in the project file for enhanced security.